Moderation: create a class to handle scheduling of infractions
The class handles application, pardoning, and expiration of infractions.
This will allow other cogs, such as superstarify, to later subclass it
to receive the same functionality that the infractions cog had.
